second commit tests sprint 7 Khamidov
