+++ src/main/java/example/ScooterUrls.java
@@ -1,0 +1,11 @@
+package example;
+
+public class ScooterUrls {
+    public static final String HOST_NAME = "http://qa-scooter.praktikum-services.ru";
+    public static final String CREATE_COURIER = "/api/v1/courier";
+    public static final String LOGIN_COURIER = "/api/v1/courier/login";
+    public static final String DELETE_COURIER = "/api/v1/courier/";
+    public static final String CREATE_ORDER = "/api/v1/orders";
+    public static final String DELETE_ORDER = "/api/v1/orders/cancel";
+    public static final String GET_ORDERS_LIST = "/api/v1/orders";
+}
