--- src/test/java/org/example/apitests/entities/requestentities/Courier.java
@@ -1,44 +1,0 @@
-package org.example.apitests.entities.requestentities;
-
-public class Courier {
-    private String login;
-    private String password;
-    private String firstName;
-
-    public Courier() { }
-
-    public Courier(String login, String password, String firstName) {
-        this.login = login;
-        this.password = password;
-        this.firstName = firstName;
-    }
-
-    public Courier(String login, String password) {
-        this.login = login;
-        this.password = password;
-    }
-
-    public String getLogin() {
-        return login;
-    }
-
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-}
