--- src/test/java/org/example/apitests/entities/responseentities/OrderResponse.java
@@ -1,102 +1,0 @@
-package org.example.apitests.entities.responseentities;
-
-import java.util.List;
-
-public class OrderResponse {
-    private Integer track;
-    private String firstName;
-    private String lastName;
-    private String address;
-    private String phone;
-    private String rentTime;
-    private String deliveryDate;
-    private String comment;
-    private List<String> scooterColor;
-
-    public OrderResponse() {
-    }
-
-    public OrderResponse(Integer track, String firstName, String lastName, String address, String phone, String rentTime, String deliveryDate, String comment, List<String> scooterColor) {
-        this.track = track;
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.address = address;
-        this.phone = phone;
-        this.rentTime = rentTime;
-        this.deliveryDate = deliveryDate;
-        this.comment = comment;
-        this.scooterColor = scooterColor;
-    }
-
-    public Integer getTrack() {
-        return track;
-    }
-
-    public void setTrack(Integer track) {
-        this.track = track;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public String getRentTime() {
-        return rentTime;
-    }
-
-    public void setRentTime(String rentTime) {
-        this.rentTime = rentTime;
-    }
-
-    public String getDeliveryDate() {
-        return deliveryDate;
-    }
-
-    public void setDeliveryDate(String deliveryDate) {
-        this.deliveryDate = deliveryDate;
-    }
-
-    public String getComment() {
-        return comment;
-    }
-
-    public void setComment(String comment) {
-        this.comment = comment;
-    }
-
-    public List<String> getScooterColor() {
-        return scooterColor;
-    }
-
-    public void setScooterColor(List<String> scooterColor) {
-        this.scooterColor = scooterColor;
-    }
-}
